Renamed button component files back to index
