+++ client/src/Components/Button/index.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Story, Meta } from '@storybook/react/types-6-0'
 
-import { Button, ButtonProps } from './Button'
+import { Button, ButtonProps } from './index'
 
 export default {
   title: 'Components/Button',
