+++ client/src/stories/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Button } from 'Components/Button/Button'
+import { Button } from 'Components/Button/index'
 import './header.css'
 
 export interface HeaderProps {
